[llc] Add a reusable LLVMLlcDriver library
